Removes generated files and adds lexer generation to the build sequence
Generated files should generally not be in VC so they are now removed.
As the lexer can now be generated automatically, I've added it to the build
process as a dependency of the parser. Note that for successful lexer
generation, the latest rexical version must be used. The git repo for rexical
is here https://github.com/tenderlove/rexical. A simple git clone, cd, rake
gem, and gem install should be all that is necessary.