Adds tests for when there can be mutliple candidates for a return type.
There can be multiple type signatures for a method that match the passed
arguments. Therefore, all the candidate types must be stored before the
call and then each return type checked against the candidate types. Adds
a test to ensure that all candidates are checked before a type error is
declared.