Adds handling of frozen types.
freeze disallows mutating instance variables, however, like Java you can
still call methods on frozen properties (provided those properties
themselves are not frozen). This means that if we have a guaranteed non-frozen
hash on an object, we can happily add or remove metadata as we please.
To implement this, the freeze method now guarantees that the _rtc_meta
property is defined on an object before the object is frozen. Further,
all metadata (including the type object) must now be stored in the rtc_meta
hash (previously the type object was cached in _rtc_type). Adds some test
cases to ensure that frozen objects work correctly.