Changes the subtype rule for dynamic type variables
Previous versions of the subtype rule for a dynamic type variable were
as follows: t <= tvar iff (t <= wrapped(tvar) and tvar is dynamic) or
 (t == wrapped(tvar) and tvar is not dynamic).
However, this won't work because then we cannot annotate Array functions
in a meaningful way.
Assume that the push method of Array is annotated as push: (t) -> Array<t>
This looks correct, however, the following perfectly valid code sequence
will fail:
a = [1,2]
a.push("foo")
as String <\= wrapped(tvar_a)
The new subtype rule distinguishes between when subtype checks are being
performed on two different type variables tvar' <= tvar. In this case,
the above subtype rules are used, except that wrapped(tvar') is used
in place of t. If the subtype rule is of the form t <= tvar, where t is
not a type variable, then the following subtype rule is used:
t <= tvar iff (tvar is dynamic) or ((t == wrapped(tvar) and tvar is not
dynamic).  Adds some test cases to enforce these constraints.