Initial handling of nested parameterized types
If a parameterized type contains as it's type parameter(s) other
parameterized types, the type parameters of those parameterized types
are unified (this process proceeds recursively). As an example, the
type of [[:foo, :bar],[:bar,:baz]] is deduced to be
Array<(Array<:foo or :bar or :baz>)> instead of Array<(Array<:foo or :bar>
or Array<:bar or :baz>)>. This process will work with further nested
parameterized types. Tests of this functionality will be forthcoming,
for now all previous unit tests are passing.