Fixes up a bug in supporting frozen objects
If an object is frozen before rtc is loaded (and thus before the frozen
hook is installed on the Object), then rtc would fail. This can occur when
code is executed in a required library file that is loaded before rtc.
This fix will simply not cache the metadata (and thus the type object)
of an object if the object is frozen. Note that this can cause problems
for parameterized types, as they will no longer be annotatable. However
I could think of no solution that did not require caching data in some
global place that would result in the type object being inellligible for
gc.  Adds a test case to ensure that this bug doesn't crop up again.