Further refines the handling of frozen types.
If metadata is not cacheable on the current object, due to the object
being frozen before rtc is loaded, then the metadata is cached instead
in a global object. Adds in an rtc_gc method (probably misleadingly named)
which will explicitly free a frozen object's metadata.  Adds test cases
to make sure that the metadata is not reinitialized for a non-cacheable
object.