Adds a lazy nominal type
The lazy nominal type relaxes the requirement that classes be forward
declared before their usage in type signatures. Instead, on use they
will try to load the class and the time of use, and if the class can
be found at the time of use, then methods will throw an exception.
Usage here is defined as follows:
* subtype checking
* method/field addition
* method/field type lookups
* attempting to access the wrapped class
* accessing or setting type parameters

Note that this is still pretty restrictive. For instance, it will be
impossible to do lazy loading for parameterized classes (as instantiating
a ParameterizeType requires reading off the type_parameters property of
the parameterized class). In addition, intersection types will not
work, as creating an intersection type requires subtype checking of
all members of the intersection type, which in turn requires subtype
checking of the arguments. To relax this further would compromise
type safety or lead to more complex problems down the line.