Adds test cases and refines lazy types
This commit modifies the lazy types so that certain operations on the
the lazy type do not cause the lazy type to be "forced" and causing an
error if the type does not exist. These methods are
* hash
* to_s
* inspect
* ==
* eql?
In addition, the equality rules for lazy types has been written.
Equality testing is complete side effect free, and as such, may
report a lazy object as being unequal to a NominalType that the lazy
type would resolve to had the lazy type been forced before the comparison.
Adds several test cases to ensure the above constraints are respected,
and to ensure that operations are properly forwards on forced lazy types.

This commit also fixes a bug in the class loader that was discovered
by one of the unit tests.
This closes issue #20.