Makes fields accessible at runtime.
Field types are now added to the type object when their type is annotated.
To lookup the type of a field, rtc_typeof is used, but to avoid ambiguity
to get the field type of the field "foo" rtc_typeof("@foo") must be used;
rtc_typeof("foo") will return the type of the getter function. This commit
also allows symbols to be used in rtc_typeof. If multiple type annotations
are put on a field, then the field has the type of the union of all those
types. Adds several test cases for the above changes.
This closes issue #2