Adds in beginnings of standard library annotations
Begins to annotate the standard library. Currently only a few methods
(those referenced in the test cases) are annotated, and only Array and
Set are annotated. Refactors the test cases to not manually set up
these classes as polymorphic. Also adds in a test case for rtc_instance_typeof
for parameterized methods.