Fixes the implementation of structural types
Fixes a bug in the subtype rule for structural types. Field types are
not checked in the subtype rule because they are not really "first class"
in Rtc's type system, and what really matters is the method interface
matches; whether those interfaces are a result of declaring a field type
or manual declaration is not important. Adds test cases for testing
structural types.