Removes unnecessary eval, and adds better error messages
On return type mismatches, the actual type of the return value is
reported. The class name of the method in which the function resides
is also reported. In addition, the eval that was used under the old
argument binding scheme is no longer necessary and has been removed
to allow for cleaner code.