Parser no longer mutates state.
When the parser encounters a production like class Array<t>, instead of
perfomring the annotation of the Array class in place, it returns a
ClassAnnotation object that the client code must handle. Refactors
what used to be in the parser into a separate class called ClassModifier.
Code that calls the parser now uses this class.
This commit closes #28.