Implements runtime type annotations
This commit allows the user to annotate the type of an object at
runtime via the rtc_annotate method. This commit also adds in subtype
checking when a constraining is performed; if a type parameter is
constrained to a type that is narrower than the current type, e.g. an
array containing strings and fixnums being annotated as Array<Fixnum>
then the newly added TypeNarrowingError is thrown. Adds in some test
cases to test this functionality.
This closes issue #9.