The method_wrapper now looks only in the class where a method is defined for a type
Previously, the type of a function was retrieved from an instance of an
object via rtc_typeof. This would act "polymorphically", so if a child
class defined it's own signature for an overriden function then that
signature would be checked against for the implementation of a method
in a parent class. This is mitigated by adding an optional parameter to
rtc_typeof called "which". The which argument must be a class object and
if specified the type signature returned is the type signature for the
given method as specified which argument. If no such definition
exists in the which class then rtc_typeof returns nil. Adds in some tests
to ensure this behavior. Field types are not yet supported using the which
parameter. Further, rtc_instance_typeof will still lookup the inheritance
chain. It may be worth adding a "include_super" parameter a la the
instance_methods defined on the Module.
This commit closes #29.