Implements class method annotations
Class method annotations may now be applied by prefixing the name of a
method with "self.", e.g. typesig("self.foo: (String) -> String")
Refactors the parser and the grammar to handle this, and adds in a
ClassMethodTypeSignature to differentiate an instance typesig from a
class typesig. rtc_type for a Class object now returns the type object
for that class' eigenclass. This allows for supporting class methods
almost completely without changes. Refactors the Annotated module to
handle class methods (there is still some opportunity for refactoring,
as there is some duplicated code). This commit also adds in support for
class fields, although their use is still untested. Test cases will be
forthcoming.
Closes #31.