Fixes a major bug in annotated.rb
Prior to this commit, the Annotated class used class variables to
maintain state. However, it turned out that these class variables
were shared among all classes using Annotated. Changes this to use
class instance variables which are not shared. Further, the method_added
hook will be invoked for instance methods added in classes that have
subclassed the class on which the method_added hook is defined. This
is now a problem as the in this case self is bound to the class in which
method has been added. Therefore we add some checks that aborts the
hook if the necessary state cannot be found.