Fixes a bug with annotating a method that has already been defined
Previously there was a bug in the code that searched to see if the
method being annotated was previously defined, which would lead to
the annotationg being added to the type object, but the method wrapper
(which implements the type checking) not being created. Fixes the bug
which now results in a huge hit in terms of performance. We probably
need caching badly.