Adding rgl to the examples.
This version is unannotated, but invoking `rake test' in examples/rgl passes.
A good test to make sure you have everything set up correctly before trying
to add annotations.

The main place we'll want to focus on annotating is probably
examples/rgl/lib/rgl/base.rb.  I'll next check in my changes to DirectedEdge
as a starting point for this.