Allows for annotating classes as not-inferrable
This commits allows for classes to be annotated as polymorphic, but
also marks them as un-inferrable. This means that classes that are
so marked can only be created via annotations. This means that an unannotated
Enumerator (for example) cannot be passed to a typed position (it will now
raise an exception).