Implements type abbreviations
Type abbreviations are now supported with the following syntax:
 type %<type_name> = <type_expr>
where <type_name> and <type_expr> are the non-terminals from the type
grammar. Type abbreviations can only be used in modules or classes that are
declared to be annotated with rtc_annotated. Type abbreviations are also
local only to the module/class in which they are declared. That is, modules
and classes do NOT inherit type abbreviations from their enclosing scope.
This is done intentionally.

Further, the hardcoded type names, %any, %none, %bool, %true, and %false
remain hard-coded because of the above restriction. No checking is in place
to ensure that these type names are not shadowed (this will be coming in a
later commit).

Closes #22.