First cut of function instantation
Functions can now be instantiated via the rtc_instantiate call on proxy
objects. rtc_instantiate takes two arguments, the first is the name of the
method, the second is a hash mapping type variables to a string representation
of the types to use for those type variables. Blocks *can* be passed to the
instantiated methods. The pseudo-proc returned from rtc_instantiate supports
call, [], and () from the Proc api, but does not support the other methods,
such as arity, parameters, curry, etc. This commit includes a first cut at
tests, but more testing is required before I call this feature totally
implemented.

In addition, in the process of creating this commit, the get_method method
of the NominalType was refactored take two paths depending on whether the
which is present or not.

Partially addresses #44.