Adds in support for autowrapping objects
This commit introduces the rtc_autowrap method that can be called after
a call to rtc_annotated. rtc_autowrap will autowrap objects created
via new in a ProxyObject. Auto-wrapping is inherited by classes deriving
from the class that called rtc_autowrap, but this can be broken by a call to
no_subtype (rtc_autowrap can be reintroduced later in the inheritance chain
if that is desired).