Fix up the HashConverter spec.
Move the specs for functions available in the block argument to
HashConverter.convert within a dsl block in that method's specs.

Avoid using global state as much as possible, and route any actual
global state through the Dsl.state hash.

The blocks for pre/post_conds and the like don't (or shouldn't) get
the block given to the method, so remove arguments that make it look
like they do.